in order to allows to change config.js from Docker volume or kubernetes volume, I move the file to config directory
On branch ocp
Changes to be committed:
	renamed:    config.example.js -> config/config.example.js
	modified:   server.js
